requirements.txt
discord.py == 2.1.0
python_dotenv >= 0.20.0
pytz >= 2022.5
googletrans == 4.0.0-rc1
emoji-country-flag
countryinfo 